requirements.txt
findspark
pandas
pyspark
yfinance
numpy
kafka-python
scikit-learn
pymongo
matplotlib
langchain
langchain-ollama
langchain-chroma
pandas
ollama>=0.1.0
httpx>=0.24.0
langchain-community
dash
plotly
pymongo
streamlit
airflow

